Fix snapshots getting deleted if followed by null snapshot (#726)
We were seeing failures after rewind where the snapshot we were trying
to delete didn't exist. We don't always send a snapshot up in
`store_chat`, so we use the previous filesystem snapshot and should not
delete it if the new snapshot is null. Full conversation here, good find
@jordanhunt22
https://convexdev.slack.com/archives/C08JFTERXUN/p1749765427053959?thread_ts=1749762232.566109&cid=C08JFTERXUN

+++ convex/messages.ts
@@ -284,7 +284,7 @@ export const updateStorageState = internalMutation({
       if (previous.storageId !== null) {
         await ctx.storage.delete(previous.storageId);
       }
-      if (previous.snapshotId && previous.snapshotId !== snapshotId) {
+      if (previous.snapshotId && previous.snapshotId !== snapshotId && snapshotId) {
         await ctx.storage.delete(previous.snapshotId);
       }
       await ctx.db.patch(previous._id, {
